Remove Project folder
