+++ public/OneSignalSDKUpdaterWorker.js
