+++ public/OneSignalSDKWorker.js
