+++ src/MakeCall/FunctionalStreamRenderer.js
